Un-hardcode host/guest IP
I was running into trouble using gitian LXC INSIDE a VirtualBox VM, because
the 10.0.2.x IP range overlaps with the range that VirtualBox uses for
its internal NAT network.

This commit makes it possible to override the host and guest IP that
is used during LXC image creation and the build, using environment variables:

- `GITIAN_HOST_IP` (default=10.0.2.2, host IP from the viewpoint of the
   VM, can be used for both LXC and KVM)
- `LXC_GUEST_IP` (default=10.0.2.5, guest IP from the viewpoint of the
   host, effective for LXC only)

When these are not defined the behavior remains the same as before.