Remove cd output from backtick output
In certain shell and environment configurations, the `cd` shell command can
output the current directory. When using backticks, this output is unexpected
and can cause errors.

Using the `Dir#chdir` block form avoids this problem.

This pattern could be refactored into a helper method to avoid repeating the
same structure.